+++ plugins/CoreHome/vue/src/index.ts
@@ -1,3 +1,6 @@
+import './AjaxHelper/AjaxHelper.adapter';
+
+export { default as AjaxHelper } from './AjaxHelper/AjaxHelper';
 export { default as activityIndicatorAdapter } from './ActivityIndicator/ActivityIndicator.adapter';
 export { default as ActivityIndicator } from './ActivityIndicator/ActivityIndicator.vue';
 export { default as translate } from './translate';
